Downloads/task/task_manager/requirements.txt
djangorestframework==3.14.0
django-filter
django-cors-headers
